Introduce a custom dialog for choosing QS tile columns/rows
This is a simple popup dialog with seekbars. The goal
here is to give the columns/rows controls it's own
view, to keep a cleaner view in main preference screens.

Also it's just nice to switch things up from time to time :P

Help fix initial values not setting properly with a
clean install.
Signed-off-by: calebcabob <calphonic@gmail.com>
Change-Id: Ia429194c61d09c9c537d25d8c08a21dcc6f5bd92